Rework to work with new Windows VMs ("1809" versions)
The new Windows VMs have WinRM and not busted SSH for remote access, so we
need to rework how we communicate with the VMs. Thankfully this turned out to
not be too bad, although it's a bit of a shame since SSH is/was nice to be
able to use.